Refactor CMakeLists.txt and stenographer.cpp
This commit refactors the CMakeLists.txt file by adding a check for the buildspec variable and reading the buildspec.json file if it is not set. It also sets the arch variable based on the platform. In stenographer.cpp, the code now includes websocketpp/config/asio_no_tls.hpp only if the platform is not Linux. Additionally, the unused parameter hdl is now marked as unused in the message handler.